Added turn off/on PID from odometer
The odomotro's three configuration modes allow power on/off or both.
The odometer shutdown function is activated when the PID reaches the set temperature (+- 0.5 C).
The number of laps of the odometer is stored in the flash every 10 minutes to prevent deterioration.
You can configure the number of minutes that the system has to be inactive to execute the shutdown function.
The lap counter can be reset from the menu by turning the encoder to the left.